Changed type of get_param_type() to return a ushort as to be consistent with t_op.

+++ vm/srcs/run/param_utils.c
@@ -6,7 +6,7 @@
 /*   By: gudemare <gudemare@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/03/09 22:14:30 by gudemare          #+#    #+#             */
-/*   Updated: 2018/03/10 02:49:48 by gudemare         ###   ########.fr       */
+/*   Updated: 2018/03/10 02:58:05 by gudemare         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,12 +18,12 @@
 ** param_nb goes from 0 to 2.
 */
 
-unsigned int	get_param_type(t_env *env, unsigned int pc,
-				unsigned int opcode, unsigned int param_nb)
+unsigned short	get_param_type(t_env *env, unsigned int pc,
+				unsigned short opcode, unsigned int param_nb)
 {
-	unsigned int	ocp;
-	unsigned int	ref_type;
-	unsigned int	ocp_type;
+	unsigned short	ocp;
+	unsigned short	ref_type;
+	unsigned short	ocp_type;
 
 	ref_type = g_op_tab[opcode - 1].param_type[param_nb];
 	if (g_op_tab[opcode - 1].ocp == false)
